+++ Firmware/Board/v3/Inc/board.h
@@ -41,6 +41,9 @@
 #define DEFAULT_BRAKE_RESISTANCE (0.47f) // [ohm]
 #endif
 
+#define DEFAULT_ERROR_PIN 0
+#define DEFAULT_MIN_DC_VOLTAGE 8.0f
+
 #define DEFAULT_GPIO_MODES \
     ODriveIntf::GPIO_MODE_DIGITAL, \
     ODriveIntf::GPIO_MODE_UART_A, \