+++ Firmware/MotorControl/oscilloscope.cpp
@@ -5,25 +5,23 @@
 #define OSCILLOSCOPE_SIZE 4096
 
 void Oscilloscope::update() {
-    // Edit these to suit your capture needs
     float trigger_data = trigger_src_ ? *trigger_src_ : 0.0f;
     float trigger_threshold = trigger_threshold_;
-    float sample_data = data_src_ ? *data_src_ : 0.0f;
+    float sample_data = data_src_ ? **data_src_ : 0.0f;
 
-    static bool ready = false;
-    static bool capturing = false;
     if (trigger_data < trigger_threshold) {
-        ready = true;
+        ready_ = true;
     }
-    if (ready && trigger_data >= trigger_threshold) {
-        capturing = true;
-        ready = false;
+    if (ready_ && trigger_data >= trigger_threshold) {
+        capturing_ = true;
+        ready_ = false;
     }
-    if (capturing) {
-        data_[pos_] = sample_data;
-        if (++pos_ >= OSCILLOSCOPE_SIZE) {
+    if (capturing_) {
+        if (pos_ < OSCILLOSCOPE_SIZE) {
+            data_[pos_++] = sample_data;
+        } else {
             pos_ = 0;
-            capturing = false;
+            capturing_ = false;
         }
     }
 }