+++ tools/odrive/utils.py
@@ -7,7 +7,6 @@ import platform
 import subprocess
 import os
 import numpy as np
-import matplotlib.pyplot as plt
 from fibre.utils import Event
 import odrive.enums
 from odrive.enums import *
@@ -49,6 +48,7 @@ def calculate_thermistor_coeffs(degree, Rload, R_25, Beta, Tmin, Tmax, plot = Fa
     fit_temps = p1(V)
 
     if plot:
+        import matplotlib.pyplot as plt
         print(fit)
         plt.plot(V, temps, label='actual')
         plt.plot(V, fit_temps, label='fit')
@@ -547,17 +547,19 @@ def dump_dma(odrv):
                      "*" if (status & 0x80000000) else " "))
 
 def dump_timing(odrv, n_samples=100, path='/tmp/timings.png'):
+    import matplotlib.pyplot as plt
+    import re
+    
     timings = []
     
     for attr in dir(odrv.task_times):
         if not attr.startswith('_'):
             timings.append((attr, getattr(odrv.task_times, attr), [], [])) # (name, obj, start_times, lengths)
-    for attr in dir(odrv.axis0.task_times):
-        if not attr.startswith('_'):
-            timings.append(('axis0.' + attr, getattr(odrv.axis0.task_times, attr), [], [])) # (name, obj, start_times, lengths)
-    for attr in dir(odrv.axis1.task_times):
-        if not attr.startswith('_'):
-            timings.append(('axis1.' + attr, getattr(odrv.axis1.task_times, attr), [], [])) # (name, obj, start_times, lengths)
+    for k in dir(odrv):
+        if re.match(r'axis[0-9]+', k):
+            for attr in dir(getattr(odrv, k).task_times):
+                if not attr.startswith('_'):
+                    timings.append((k + '.' + attr, getattr(getattr(odrv, k).task_times, attr), [], [])) # (name, obj, start_times, lengths)
 
     # Take a couple of samples
     print("sampling...")
@@ -574,7 +576,7 @@ def dump_timing(odrv, n_samples=100, path='/tmp/timings.png'):
 
     plt.rcParams['figure.figsize'] = 21, 9
     plt.figure()
-    plt.grid('both')
+    plt.grid(True)
     plt.barh(
         [-i for i in range(len(timings))], # y positions
         [np.mean(lengths) for name, obj, start_times, lengths in timings], # lengths